Early work towards HW scroll + sprite erasers!
This is a pretty hacky commit which changes the engine from
rendering to a 320x200 buffer to rendering to a larger buffer
and scrolling within it. We'll later use this for HW scroll.

It also enables Sprite Erasers. These are of no use yet,
but when the tile refresh code is finished, they'll give us
something of a performance boost. (Plus, that's the way the original
games did it, so we must implement it!)