Added high score entry and display
There is a bizarre bug with entering shifted letters.  If the game is run normally
then shifted letters cannot be entered at the high scores screen (only lower case
letters are printed).  However, if the debugger is set to break at id_in.c : 330 (or possibly
some other place), and shift is pressed (to activate the breakpoint) and then the shifted
key is pressed WHILE the game is paused, then a capital letter will be printed.