Make F11 toggle fullscreen.
This adds the VL_ToggleFullscreen() and VL_ToggleAspect() functions,
which can toggle fullscreen and aspect correction respectively, while
the game is running. It adds a new function to the VL backend interface,
'flushParams', which tells the interface to use the new values of vl_isFullscreen
and vl_isAspectCorrected immediately.

I selected F11, as Alt+Enter is more likely to interfere with gameplay (Pogo/Status),
but this can be changed/made configurable/etc.