Initial work towards doing dirty-rectangle drawing
This adds a whole bunch of (still broken) code to only update
areas of the screen which actually change. The bugs here largely
relate to the fact that we manually copy the scrolled tiles in the
surface, rather than wrapping as the original does. The fact that we
do sprite shift emulation inside ID_VH is also a problem, as it means
we don't know exactly where to erase things.