More dirty-rectange fixes. It sort-of works now.
This adds a few more fixes for the dirty-rectange drawing. The game
is playable with it, now, though it's still very hacky and there are
bad interactions with menus and the like (need VWB_ functions).

There are still a number of hacks, so it's not as efficient as it should be:
- We add a 8px border around every sprite to account for potential shift differences.
- We redraw the entire screen on every tile scrolled, as we don't scroll the frontbuffer.
- Generally, we're doing and undoing a lot of work when rendering sprites due to the
  decision to put a lot of the work into id_vh.