Super-hacky fix for the arrow in the end text being two pixels off
The x coordinate for the "next page" arrow in the ending text is not divisible
by 8. As bitmaps are always aligned to an 8px (byte) boundary, this means that
the arrow actually renders two pixels further left. We correct for that here by
and-ing the value with ~3, which makes the arrow no-longer overdraw the screen
border.

This code should be nicer, and I apologise.