Cleanup the Terminator code a bit, and fix a memory OOB bug.
This generally cleans up much of the Terminator code. I'm unsure if it would be
a good idea to totally get rid of the shifting and switching to using runs of
ScreenRect_PM (like the Zoom Out code does) would be better than converting to
1bpp and then undoing it in id_vl. I've left it for now.

Most of this was in order to find an out-of-bounds read from the shifted
commander bitmap. This turns out to have been a result of the middle case in
AnimateTerminator (clipped on both sides) being inconsistent about using 320/8
bytes and setting cmdrWords to 21 (for 336/8 bytes). Once fixed, we're back to
being (to a cursory glance) free from such memory errors.