Add this Keen 6 specific code piece.
In levels 6-8, there are cases in which walls meet sloped floors.
Without the added code, if Keen is pushed with right/left key towards
a wall while standing on a sloped floor, then glitches are reproduced.
Almost-identical glitches can be seen in Keen 6 v1.0, and this
was fixed for v1.4. Note that this code is DISABLED in Keen 4-5,
so problems are expected in custom Keen 4-5 levels
with such wall-floor meetings.