Get the Star Wars scroller working.
Notably, this also adds some features to ID_VL for changing the surface used
as the "screen", and a "PGet" function for getting the value of a pixel in
any surface. We use these to manually scale the rows of the Star Wars text one
pixel at a time, rather than trying to compile scaling functions.