Support ?EGAHEAD files without the terminating entry.
Some mods have an EGAHEAD file which does not have the size of the
corresponding EGAGRAPH file as the final entry. This causes omnispeak
to crash trying to load the final chunk.

This change detects attempts to load the final chunk in such files,
and uses the length of the ?GAGRAPH file present. This fixes the
crash.

"The Eight Accumulators" High Scores demo is a good example of such
a crash.