Add a memory arena class to allocate action strings
This adds a new feature to id_mm, the arena. The arena allows quick allocation
of memory which will all be freed at once. It's used here for strings in the
action parser.

Indeed, there are two arenas used there: one which is used for temporary
allocations (and is reset after every action is loaded), and a larger one
which stores all of the strings for action names. This latter one is 16K.

There are two provided implementations of the arena: one is a simple
mark-release allocator: it allocates a single chunk of memory which is
partitioned into used and free areas by a simple offset. It's extremely
quick. The other "debug" implementation adds some extra tracking around the
system malloc. It's much slower, and allocates a bunch of extra memory for
this tracking info, but it allows tools such as valgrind to know how the
memory is layed out. Enable it with the ID_MM_DEBUGARENA define.

Note that the STR_Pool/STR_UnPool functions have been removed.