Properly cache sprite lumps
This fixes the sprite "lump" caching for Keens 4 & 6, and adds it to Keen 5.
In the process, it guards all of the CacheGrChunk calls in IF_RF with warning
messages. I've played the games sucessfully with these as Quit()-able errors,
but it's probably best to have this nice fallback for now.

There are some spurious warnings printed in Keen 6's Guard Post 3, which has
a NULL background tile.