Super-dodgy Joystick support.
This adds basic support for Joysticks. It notably does not add support for the
SDL gamecontroller features, nor does it add the calibration code. Up to two
joysticks attached to the system will be selected semi-randomly as Joystick #1
and Joystick #2 in the options menu.

The game can be played with the joystick, but not all of the menus work (and,
when they do, they can have weird speed issues).

Note that at the moment the actual value from the joystick is discarded and
converted to a single IN_Motion value. The original game does this for the
actual game, but menu code sometimes made use of the actual x/y values.