Add support for Keen 6 v1.4 EGA (thanks NY00123)
This adds the extracted data from Keen6 v1.4 EGA (thanks to NY00123), and adds
some basic autodetection code to the game so that it can support both.

The autodetection code is kind-of hacky: there are two EpisodeDef structs, one
per version, and the isPresent method will set a ck6_episode pointer to the
one with an EGAGRAPH.CK6 file which matches the right size.

You can choose which version will have its data copied to the bin directory with
the KEEN6VER makefile parameter.