Upload my super-hacky, super-unfinished Vulkan backend.
It doesn't work. It will display a few things on Linux if you compile it in,
and it will then crash if you try (for example) to resize the window.

The code is ugly and incomplete, but it might prove interesting to someone,
and maybe I'll (or you'll) finish it at some point.