Rework the end text help to use IN_UserInput()
The current end text uses repeated calls to IN_CheckAck() and VL_GetTics
to check for input while making the arrow flash. By using IN_UserInput()
instead, which was designed for this sort of thing, we no-longer end up
doing repeated unnecessary calls to SDL_Delay() and similar, which cause
this screen to be very unresponsive on my machine.