Isolate some SDL dependencies, add IN/SD backends
This is a significant change which is the initial work in separating out
the parts of Omnispeak which depend on SDL, so non-SDL backends can be
written. It also, by splitting out the ID_SD and ID_IN subsystems into
frontend and backend parts, makes room for new backends. One such
example could be the use of different OPL2 emulators, or OPL2
passthrough.

Some notes on the implementation so far:
- A number of uses of the SDL endianness functions have been replaced
with CK_Cross functions.
- The SD subsystem has a new pluggable backend system. The backend
receives raw PkSpkOn and alOut requests, so must emulate a PC speaker
and OPL2 chip.
- There is still a bit of SDL stuff in the SD frontend, namely a use of
SDL_GetTicks in SD_GetTimeCount()
- A few other cleanups and bugfixes to SD, partularly around locking the
audio thread.
- The IN subsystem has a new pluggable backend system. It mostly just
forwards raw ScanCodes to the frontend, and has some joystick support.
It needs a lot more work for when we decide to handle joysticks
properly.
- There is a new VL_Yield() function to replace the random calls to
SDL_Delay(1) around the place.
- Probably lots of bugs and problems. Yay!