DOS/DPMI: VGA 13h → EGA 0Dh
This is some early work towards the DOS/DPMI version using the proper
EGA video mode. It's actually much buggier than the old VGA version, but
also (when in-game) fast enough to be playable.

Here are the major changes:
- The VL backend interface now supports a ScrollSurface function, which
scrolls a surface by some amount. This is used to implement the
smooth-scrolling hack.
- The VL backend interface now also supports double-buffering. Only the
DOS backend uses it, as it breaks a lot of things. I'm still trying to
work out how best to handle this, as the original only uses
double-buffering in-game, and things like the terminator scroller seem
to need more memory than is available if double buffering is enabled.
(One possibility would be to have single buffering in VL, and implement
double buffering by changing the scroll offset a lot.)
- ID_RF now supports double-buffering with ALWAYS_REDRAW disabled.
- ID_RF now supports scrolling using VL_ScrollSurface, and scrolls the
dirty block buffer as well.
- Added a bunch of min/max functions which I need to clean up. The
CK_Cross versions are the canonical ones I'll make everything use.

The major bugs:
- Double buffering breaks everything UI-wise, makingt them flicker
uncontrollably.
- TerminatorScroller doesn't work in DOS
- SWScroller doesn't work in DOS
- As we don't have sprite shifting, it's hacked in ID_VL_DOS, and breaks
when clipped on the left-hand side of the screen.
- Lots of random memory corruption (particularly in CK_INTER stuff)