Fix a couple of issues related to the Wetsuit
CapnClever reported a few inaccuracies with the Wetsuit in issue #7.
This fixes two of them:

1. The wetsuit in Miragia is only supposed to be collectable when
touching the ground (topTI is set), but we had the opposite. Whoops.
2. There are a number of problems with the resulting dialog. This fixes
the vertical alignment, which should add 12px instead of 6px.

The other issues (around the sound/etc changes) are a little more
interesting and will be fixed in a future patch.