Text shouldn't disappear in Wetsuit message
The text in the Wetsuit message was disappearing when redrawing the
portrait of Keen. This was because we were re-drawing the window
(removing its contents). This, in turn, was so that US_WindowW() was
being reset, to work around our (incorrect) re-subtraction of 0x30 from
it.

This has been updated to match the original game, which doesn't call
US_CenterWindow a second time, nor subtract 30h from US_WindowW during
the second VH/VW_DrawBitmap call.

This fixes the 2nd issue in #7 (though there are still some timing
issues associated with the CK_ShowWetsuitMessage()).