Support Demo Recording (F10+D)
This commit adds support for recording demos with the F10+D cheat.
Support is added in ID_IN (specifically IN_ReadControls()) to keep a
copy of any control data read, as well as some new functions for
initialising this and saving to a file.

This commit also fixes joystick buttons not being returned properly.
Oops!