Variable parser improvements, including string concatenation
This improves the lexer in ID_STR to support an explicit STR_Token type,
which allows things such as the type of token to be returned. It also
allows the end-of-file condition to be explicitly supported.

Many strings in Keen (such as, for instance, the story text) are
multiple lines long, and benefit from being split up into multiple
concatenated strings in the STRINGS.CKx file. The "%string" command now
accepts multiple strings as arguments, which are concatenated together.
(The maximum length is hardcoded at 4096 bytes at the moment, and it
gets crashy if that's exceeded, so be warned!)

Finally, support has been added for an "%include" command to include
another variable file.