Add an SD backend for Raspberry PI/OPL2
This adds support for using the ArduinoOPL2 board to drive an OPL2 from
Raspberry PI GPIO pins.

Note that this needs to be built manually (by changing the makefile to
use the id_sd_opl2.o module instead of SDL2, and doesn't support the PC
speaker at all: no linux-level sound output exists, both music and sfx
are output from the OPL2.

This necessitated a new timing system (as the previous ones used either
the SDL audio callback or PIT (under DOS) for game timing. This one uses
an SDL thread (note that using SDL_AddTimer doesn't work, as it only has
millisecond resolution, and we need to update at 560Hz, so the game is
noticeably slower if we run at 500Hz).

See:
https://github.com/DhrBaksteen/ArduinoOPL2

(We don't actually use the OPL2 library, which is C++, but call wiringPi
directly. The pins used are hardcoded, but should be the same as with
the OPL2 library. Note, too, the the Pi will need SPI enabled.)