Use GL_LUMINANCE8 instead of GL_R8 (OpenGL compat)
We've been using GL_RED/GL_R8 as the texture format in the OpenGL
backend (because, if I recall, it fixed some issues with older drivers),
but it was actually only introduced in OpenGL 3.0. Since we claim to
support OpenGL 2.0, we should use GL_LUMINANCE/GL_LUMINANCE8 instead,
which is better supported.

This fixes, for example, the "experimental" Raspberry Pi/Raspbian OpenGL
driver (which now works very well with Omnispeak).