Properly calculate integer backbuffer sizes in SDL backends (fixes #13)
We weren't taking the size of the overscan border into account when
calcuating vl_integer{Width,Height} in VL_CalculateRenderRegions. This
was resulting in the integer rendertarget being a multiple of 320x200,
but including the overscan borders. Some ugly scaling artefacts ensued,
as we were doing nearest-neighbour scaling at an almost-right
resolution.

This fix replaces the use of VL_EGAVGA_GFX_{WIDTH,HEIGHT} with
VL_VGA_GFX_SCALED_{WIDTH,HEIGHT}_PLUS_BORDER in this calculation. This
actually results in us always scaling the height by at least 2 (as the
overscan border is an odd number of scanlines wide, and pixels are 2
scanlines each). This may be slightly less efficient, but is required to
prevent this issue in all cases.