FullClipToWalls: If episode is not set, all obclasses are bad
In CK_PhysFullClipToWalls, we delX and delY are unitiialised if the
episode is not set. Given that for unknown objclasses, we normally
Quit(), make the default case fall through and do so here (if the
episode is unknown, so are all of the sprites in it).

This was found by clang's static analyser.