Add timer-based fallback for ID_SD_SDL (and /NOAUDIOSYNC)
Game (and sound) timing for Commander Keen traditionally was determined
using the PIT (Programmable Interrupt Timer), which causes the
SDL_t0Service callback to be called at a regular rate.

For Omnispeak's ID_SD_SDL backend, the number of sound samples generated
is used to drive this timer: this is high-resolution, and guarantees
that the sound will have accurate timing (useful if the game is being
recorded).

However, if the SDL audio device can't be started, or if it's broken,
timing for the whole game will fail. This change adds a fallback using a
separate timing thread, which will drive the timer independently of the
rest of the sound system. It is used if the SDL audio device can't be
opened (where otherwise, the game would hang as no time ever passes),
and can be forced with the /NOAUDIOSYNC parameter, if there are issues
with the audio API in use. (I've had some issues with wine and SDL's
wasapi driver, and this makes the game playable, though the sound is
still very broken.)

Note that SDL1 lacks the SDL_GetPerformance{Counter,Frequency}()
functions, and so we fall back to SDL_GetTicks(), which lacks precision.
Indeed, the precision is low enough, that the time is rounded down,
causing the game to run noticeably slowly. A warning is printed if this
fallback is used with SDL1.