Some small fixes for /NOAUDIOSYNC
This makes two changes for the /NOAUDIOSYNC SDL audio timer fallback:

1. It keeps track of the next tick time independently of the current
timer, preventing the rounding errors from accumulating and making SDL
1.2 slow. It's still low-precision, and so there'll be some timer
judder, but it's very minor now.

2. Properly initialise some of the timing variables, as the game doesn't
initialise them until after the audio callback started. This could cause
the audio callback to wait forever, deadlocking the game. This also
seems to make the wine wasapi implementation a bit happier, as it seems
very sensitive to spending too much time in the audio callback.

Ultimately, it'd be nice to fix a few more things. It'd be nice to
finally move the OPL2 emulation out of the audio callback, as well as to
get rid of the 'volatile' variables (I wanted to use atomics, but SDL
1.2 doesn't support them).

Still, this should be an improvement.