Various fixes for DOS
This change adds a number of fixes for the DOS backend. Most notably, it
adds a singleBuffered argument to the ID_VL "present" interface: when
true, the buffer is not flipped on present (though scroll coordinates
and the like are updated). To support this, ID_RF now calls VL_Present
itself, after triggering a page flip with the new VL_SwapOnNextPresent()
function.

Also changing are:

1. Palette setting in VL_DOS is now fixed. Instead of using the VGA
palette registers, the BIOS "SET ALL PALETTE REGISTERS" call is used
(Int 10h, AX=1002h), which works properly in 16-colour modes. This
requires the palette be stored in BIOS-accesible memory (and with the
intensity bit being the green intensity bit), so DPMI is used to
allocate some DOS memory, and the converted palette is loaded there.

2. Several bugs in the plane-masked versions of the DOS graphics
functions were fixed, where the mask was being applied to the input
colour rather to the actual blit, causing some planes to be incorrectly
overwritten

3. VL_DOS_SurfacePGet() now calculates the bit offset correctly, fixing
the Star Wars scroller text.

4. VL_DOS_BitToSurface{,_PM}() now properly handle non 8-px-aligned
offsets, fixing some juddering in the Terminator scroller. (Note that
there still can be some juddering due to tearing, but it's much better
than it was.)

5. A number of variables where were mistakenly named "null" have been
renamed to "dos".