added experimental quicksave/quickload feature
Currently bound to F5 (save) and F9 (load) keys.

Always saves to and loads from the last (sixth) save slot;
Quicksave overwrites it without asking.

The saving and loading functions in ck_us_2 have been
refactored into a "shell" part and a "main" part,
the latter of which is used for quicksave/quickload too.
Not sure whether US is the correct subsystem for these
though; maybe they should be moved to another one.

Also extends the pause feature a bit to show
arbitrary messages, used here to show "GAME SAVED"
and "GAME LOADED".

Possible issues detected while using this feature
in a playthrough of all three episodes (not
necessarily related to quicksave/quickload, but
not possible to rule it out either):
- lots of "uncached graphics chunks" warnings and
  crashes happen when using this; not sure if it's
  directly related to quickloading (as it doesn't
  do anything different from loading via the menu)
  or if it would also happen when loading/saving
  frequently using the menu; at the very least,
  quicksave/quickload makes saving and loading so
  easy that it's used more often and the issues
  creep up much faster
- sometimes the camera position is going a little
  crazy, showing out-of-map tiles and crashing the
  game (happened once in E4 in Crystalus in the
  final room, but couldn't reproduce since)
- end text in ep.5 couldn't be loaded properly
  once (^P check failed), but failed to reproduce