fix audio glitch when loading
When loading a game (either from menu or quickload),
there was a short audio glitch because the music
started, managed to play a first note, but stopped
immediately after that because the game is paused.

This little fix makes the music start in "suspended"
state iff the game is currently paused. Playback
resumes (or, in this case, *starts*) normally
when pause mode is left.