moved joystick quantization code into id_in.c
All meaningful joystick quantization (deadzones,
segments) was done in the SDL backend so far.
This didn't matter much, as the SDL backend is
the only one with joystick support anyway, but
to make the code more generic and extensible,
it has now been moved into the common code
in id_in.c.

Fallout of this change:
- the backends don't have any business in
  dead zones any longer, the joySetDeadzone()
  method vanished
- the backends need to provide information
  about the range of the axis values that
  are produced by joyGetAbs() though
  (this is required for deadzone computation,
  and also helps with backends that may use
  unsigned instead of signed output coordinates)
- the old deadzone-only quantizer and the new
  segment-based quantizer are selectable at
  runtime, because why not?
- the quantization code was identical for
  the menu (IN_ReadCursor) and gameplay
  (IN_ReadControls), so it has been extracted
  into a common function (IN_GetJoyMotion)