Add support for building a VANILLA build
This adds a new config header: ck_config.h which contains #defines for
various features and options.

Notably, it automatically enabled/disables some of these depending on if
we're building a DEBUG version (e.g. the debug Arena allocator), and if
we're build a VANILLA version (which disables omnispeak-specific
features such as quicksave and the new graphics options).

The Makefile has been updated to support these options, and the new
Joystick options and Quicksave have been put behind #defines to mark
them as omnispeak-specific.

Going forward, user-visible changes should go behind a #define like this
and be excluded from VANILLA builds. I'll also (at some point) go back
and add some more options to ck_config.h and maybe put some other things
behind the VANILLA flag.

(NOTE: There are also a few bits of code reformatting, and a couple of
changes to make clang-format not puke. Most of these are related to this
change, but there are a couple of others scattered around.)