Fix and enable dirty-rectangle drawing.
This (large) change fixes most of the issues with the more optimised
dirty-rectangle drawing feature, allowing us to disable ALWAYS_REDRAW by
default. Only areas of the screen which have actually changed will be
redrawn.

In particular, the following things have changed:
1. The VHB_ functions have been improved, and are now used everywhere
they should be. This required fixing the scroll offset (we & the
x-coordinate with 8 as ID_RF scrolls within a 16px region, but the VHB_
code should be drawing into an 8px — i.e., 1 byte in EGA — aligned
region), as well as adding the missing VHB_DrawSprite() function.
The ID_US code has been updated to use the VHB_ functions, as has the
help text code in CK_TEXT, and most of the code that draws dialog boxes
(e.g, the Princess Lindsay, etc, messages and the debug windows like the
Sprite Test).

2. A number of fixes were made to the Status Window. It now uses the
VHB_ functions, and the size of ck_statusSurface has been increased to
take possible scrolling into account. The uncovered areas are now
re-drawn: the top and left from the ck_statusSurface, and the screen
underneath from the new ck_backupSurface, which has the old contents of
the screen copied to it. In the process, the calculations for the
positions have been cleaned up a bit, fixing an issue where the
top/right bitmap wasn't scrolling in correctly.

3. ALWAYS_REDRAW is no-longer enabled by default in ck_config.h

There are still some issues with double-buffered backends, like DOS and
SDL2GL when NUM_BUFFERS is increased, but they are mostly due to the
lack of VL_SyncPages() or similar: the previous page is briefly shown
when windows appear in some cases. We also rarely call something
equivalent to VW_UpdateScreen() to explicitly redraw from the update
buffer, or RF_FixOffs() to restore the scroll position after we have to
break it in some cases (help screens, us_2 cards, etc). Neither of those
seem to be causing problems.

There are also a few cases where, by using the VHB_ functions on an
offscreen buffer (like the status screen), that must be marking a bunch
of tiles in need of redraw needlessly.

Finally, for double-buffered backends, we shouldn't need the separate
ck_backupSurface: we could just use the other page (after
VL_SyncPages()). This is what the original game does, and would be
faster and more memory efficient (especially on DOS, where the other
page would be in EGA memory and could be blitted to/from with all pages
enabled at once.)