Properly shift sprites.
Previously, we were "emulating" sprite shifts by rounding the
coordinates in VH_DrawSprite(), and then calling into VL with
non-byte-aligned coordinates. This was quite sensible for the PAL8 based
backends, but meant that the DOS backend was (incorrectly) trying to
shift and clip sprites every time they were drawn.

This change implements CAL_CacheSprite() and CAL_ShiftSprite(), meaning
that sprites are shifted when cached, rather than when drawn. The
pointer in the ca_graphChunks[] array points to a VH_ShiftedSprite
structure, which has enough room for four shifts.

There are also some fixes in ID_VL_DOS, but it still needs a lot of
work. It's _mostly_ playable now, though, modulo menus and crashes.

In the future, it'd be nice to make this a bit more optional (it's
wasted work for the PAL8 backends, though given it's accurate to the
original, I don't mind it being there), and maybe allow VL backends to
suggest their own "shifted" versions of sprites (as a surface type, for
example), which could be something like a pre-converted-to-PAL8 version.
(That's what I did for the Steam Keen Dreams port.)

But for now, this is fine. It should be a no-op from a user point of
view, but there may be bugs which break things.