Fix the Star Wars scroller with "buffered" drawing
When we changed over ID_US to use VHB_* functions everywhere, the Star
Wars scroller broke, only displaying the first few lines of text. This
is because the Star Wars story is drawn to the screen using US_CPrint(),
which in turn calls down to VHB_DrawPropString(). As a VHB_ function, it
needs to update the dirty-block buffer, which is only screen-sized, so
the offscreen surface which holds the Star Wars text is now limited in
height to one-screen, which is insufficient.

To fix this, the original game alters the screen offset after every
line, meaning that the text currently being drawn is always on-screen.
We don't allow such a thing in Omnispeak at the moment — there's no way
of changing a surface's offset — so instead we now draw with the
VH_DrawPropString() function to bypass the dirty-block buffer.

The original game actually has a US_SetPrintRoutines() function (which
is unused) which would allow us to override the print function here.
This change adds support for that (making it a bit user-friendly by
allowing "NULL" to represent the default functions), and uses it in
CK_DrawSWText().

All now seems well.