Test builds on Linux on each pull request / push
As mentioned in issue #19, it'd be nice to have some sort of CI system
with which to confirm that pull requests and pushes actually build
correctly. This is a first step towards doing this, using GitHub actions
to build Omnispeak on the latest Ubuntu, with all options set to
default.

If this works, I'll look into extending it to build the other versions,
which is where it's likely to cause the most problems. But this simple
test should be better than nothing.

Ultimately, we'll want to actually test the compiled binary, using eith
the id_{cl,in,sd}_null backends or something like the SDL software
rasteriser. One thought is to have a set of demos and dumps which we can
easily compare.