improved joystick motion in menus
Navigating menus with the joystick used to be very slow
(at least on backends which use VSync in the menu).
This patch ensures that when pushing the controller up
or down from the neutral position, there is always a
change in the selected menu item, making menu navigation
much quicker.