allow returning to menu with a joystick button
This adds a fourth mappable button for returning
to the menu (a.k.a. wrist watch), i.e. typically
the Select button.

To make room for this, the deadzone is no longer
stored in the configuration files: we only
have space for four joystick-related configuration
items, and something's got to give.