Fix misaligned high scores (VHB->VH)
This is (like the issue with the Star Wars scroller being cut off[1]), a
bug due to VHB_ functions being used with US_Print().

In this case, the issue is that the text is re-aligned (depending on the
scrolling), to fix 8px boundaries (so that any Tile8s in windows are
aligned).

We fix this the same way, by using US_SetPrintRoutines() to revert to
the non-buffered version of the print functions.

[1]: https://github.com/sulix/omnispeak/commit/0e4e2403136289615c14312782b1f25c3e5924ab