Add integer scaling support
Currently, Omnispeak always tries to fill the screen (possibly with some
aspect ratio correction), typically using a two stage nearest-neighbour
then bilinear process to try to preserve pixel shape.

Alas, this has a couple of downsides when the resolution is not an
integer multiple of the original game resolution:
- Pixel edges are blurry, as they're still being filtered.
- On renderers without bilinear filtering support (e.g, the SDL2
software renderer), pixel sizes are inconsistant, which is ugly.

This change adds a new "integer scaling" mode, which will scale the
screen content to the largest integer multiple of the actual screen size
(320×200 if the overscan border is enabled, 336×208 if enabled) which is
less than the renderer's output size. (A black border will fill the rest
of the screen).

This can be enabled in the menu, or with the /integer command line
option.

Note that the overscan border size is changed slightly when integer
scaling is enabled: the top and bottom borders are now 8 scanlines
instead of seven, as we'd otherwise not have pixels line up on integer
boundaries if the top/bottom borders were 3½ pixels each.