Optimise the SDL2 renderer for integer scaling
When we're using integer scaling, there's no need to have the two-stage
scaling process, as the bilinear stage should be a no-op.

This change bypasses the scaled render target when integer scaling is
enabled. It also prints a warning if the software renderer is used and
integer scaling is not enabled.

Finally, it no longer requests only accellerated render drivers:
hopefully this means SDL will fall back to the software renderer more
readily.