Add NOCOPY and NOAUDIOSYNC options to the README (fixes #23)
The NOCOPY (disable the Keen 6 copy protection) and NOAUDIOSYNC (ID_SD
timing is based on a timer thread rather than the audio callback in
id_sd_sdl) were not documented anywhere. This adds a brief description
of them to the README file.

Ultimately, as we add more command-line options, it probably would make
more sense to move some of these to a separate bit of documentation
(particularly for things which are backend-specific). Also, there's a
possibility /NOAUDIOSYNC will become the default, and the other path
will go away.