Rewrite the ID_SD implementation (but not backends)
This is the first in a series of patches to improve the ID_SD (Sound
Manager) implementation. It consists of a complete rewrite of
id_sd.c/id_sd.h, which should be compatible with the original, but with
somewhat cleaner code. It's still compatible with the existing code and
backends, which may get some cleaning up later.

Note that, being a rewrite, it may be a tiny bit less accurate / more
buggy. Please report any such issues: the aim is for it to be just as
accurate as the original, but to be a clean rewrite which will be easier
to work with (particularly as a couple of new features are added later
in the series).