id_sd_dos: Properly delay in between writing to OPL2 registers
I finally tried this on a machine with a real SB16, and the most
horrible noise resulted. Sure enough, the problem is that I wasn't
giving the appropriate delay between writes to the OPL2. While that's
fine on DOSBox (and, it turns out, the ESS chip I was testing on), it's
a no-go on the old 486 I grew up with.

This fixes it, by using the recommended waits.

Note that we're keeping interrupts disabled for the entire wait at the
moment: Keen re-enables them immediately after the data write.