id_rf: id_vh: Make SpriteDrawEntries use pre-shifted sprites
There was a small graphical glitch which would occur because the sprite
position, the shifted sprite graphic as rendered, and the sprite eraser
didn't quite match. This resulted in, for example, the bottom of the
pogo stick flickering through foreground tiles.

To fix this, we want to do what the original game did, which is to shift
the sprites on placement (in RF_PlaceSprite/RF_AddSpriteDraw), and then
just blit the shifted data on rendering. The x/y/w/h on the
SpriteDrawEntry then outlines the sprite as rendered — the origin and
shifting already applied — allowing it to be erased properly by just
copying that bounding rectangle.

This also makes sure that the dirtied blocks are correctly applied to
the areas of the screen actually dirtied.

To implement this, instead of storing a pointer directly to the EGA data
in the SpriteDrawEntry, we just also store the shift, and have
VH_DrawShiftedSprite() functions which accept a shift number and render
that shift.