id_fs: Introduce a new filesystem manager
This change creates the ID_FS subsystem, which handles all filesystem
operations. It steals some code from ID_CA and CK_Cross, and introduces
the new FS_File type (currently just a typedef of FILE*), and functions
for performing I/O on it.

The main conceptual change here is that files are now split into three
different types:
- Keen files, which are bundled with the original game
- Omnispeak files, which are bundled with omnispeak (these are the
  ACTION.CKx files, and the various extracted headers and huffman
  dictionaries).
- User files, which are modified by the game, and consist of config
  files and savegames.

These can all now be in separate directories, rather than all files
needing to be in the current directory. This means that, for example,
the game data could be on a read-only filesystem, with the savegames
stored elsewhere.

A few other things to note:
- Everything defaults to using the current directory, and the code for
  detecting episodes assumes that the paths are still set to ".".
- Omnispeak files are actually looked for in the Keen directory first:
  this allows mods to override them.
- The Keen path can be overridden with the /GAMEPATH option.
- The User path can be overridden with the /USERPATH option.
- There are three implementations of the file I/O: for Linux (using
  openat()), for Windows, and for "other", which is for other, vaguely
  POSIX-compliant systems, including DOS. All except Windows work on
  case-sensitive filesystems.
- There's very little error checking thus far, and a few small
  inefficiencies here and there, so you can crash it pretty easily by
  passing invalid paths.

More to come!