ck_quit: Show a messagebox on error under SDL
If Quit() is called with an error message, and a version of SDL which
supports message boxes is used, display the error message as a message
box.

This should be more user-friendly on GUI systems than requiring people
to look at console output to find out why the game's crashed. (This is
also something we did in the KDreams Steam port)