ck_play: Don't Present in CK_ScrollStatusWindow()
The Status ("Enter") screen was much choppier scrolling in and out than
it should be. This was because CK_ScrollStatusWindow() was calling
VL_Present() every frame, even though it is called from RF_Refresh(),
which does its own VL_Present(). This was originally there because
CK_ScrollStatusWindow() called VW_UpdateScreen(), which would update any
dirty blocks, but that doesn't seem to be required here.

With this fixed, the status window now scrolls in and out smoothly.