Makefile: Pass -mno-ms-bitfields again to workaround old mingw32 bugs
In Pull Request #18 (specifically commit 2e57635 ), we removed the
-mno-ms-bitfields option, as it didn't seem to be doing anything: none
of the structures we cared about used bitfields.

However, this may have caused the CI Windows builds to no-longer be able
to load levels, as they use an older build of mingw (specifically a GCC
6.3.0-based one), which is affected by a bug with packed structures:
https://gcc.gnu.org/bugzilla/show_bug.cgi?id=52991

Because the CA_MapHead struct is packed, and is filled with an array of
uint32_t which is unaligned by definition (following the 2-byte rleTag),
it falls afoul of this bug, causing the wrong map header offset to be
read, a bad seek to occur, and the game to Quit with "Couldn't read map
header from GAMEMAPS!"

This bug was reported by @tgies in issue #28.