id_ti: Handle out-of-bounds tileinfo accesses (Issue #28)
In some circumstances, when using a glitch to read past the bounds of a
level, invalid tileinfo data can be read. This reads potentially invalid
memory, and can therefore cause a crash.

With this change, all out-of-bounds tileinfo access returns 0. The
glitches still appear to function as expected in DOS with this change,
but without crashing. See Issue #28 for more details.

Note that in debug builds, we log a warning for each out-of-bounds
tileinfo access. This spams a lot of log messages in the case of this
glitch (when level 2 is played after level 1), so may slow the game down
a bit.