id_vl_dos: Fix clipping issue w/large sprites off top of screen
There's been a longstanding bug in the way the DOS VL backend uses the
VL_Clip() function to clip data before drawing. The scanline (_y) loop
in all functions which used it was looping between initial_y and
final_h, which assumed final_h was the bottommost y coodinate, rather
than the width. The loop bounds should've been
[initial_y,initial_y+finalw).

This only ever really affected sprites, as they're the only things which
are actually clipped in ID_RF. In particular, it could be seen in Keen
5, level 4: drop down below the Robo Red, and position Keen on the pole
such that the top of the Robo Red is offscreen. Inexplicably, the bottom
of Robo Red will also be cut off (by however many tiles of Robo Red were
offscreen).

Also, add a small bit of documentation for VL_Clip.